S3UTILS-238: move configureCrr/removeCrrConfiguration into S3Setup.js
Lift the CRR setup and teardown helpers out of the
listObjectsByReplicationStatus functional test and into
tests/utils/S3Setup.js so they can be reused by other functional tests.
configureCrr gains an optional `storageClass` option for callers whose
flow matches by destination storage class.
